+++ Runtime/Settings/PushTokenService.cs
@@ -1,0 +1,23 @@
+﻿#nullable enable
+
+namespace AffiseAttributionLib.Settings
+{
+    public enum PushTokenService
+    {
+        FIREBASE,
+        APPLE
+    }
+    
+    public static class PushTokenServiceExt
+    {
+        public static string? Service(this PushTokenService service)
+        {
+            return service switch
+            {
+                PushTokenService.FIREBASE => "fms",
+                PushTokenService.APPLE => "apns",
+                _ => null
+            };
+        }
+    }
+}