+++ Runtime/Usecase/IPushTokenUseCase.cs
@@ -1,0 +1,13 @@
+﻿#nullable enable
+using AffiseAttributionLib.Settings;
+
+namespace AffiseAttributionLib.Usecase
+{
+    public interface IPushTokenUseCase
+    {
+
+        void AddPushToken(string pushToken, PushTokenService service);
+        string? GetPushToken();
+        string? GetPushTokenService();
+    }
+}